contracts: move mocks
